Split domain APIs to libvirtApi/domain.js
